[TWINFACES-479] fix: resolve PR comments
