+++ src/app/workspace/user-groups/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { UserGroups } from "@/screens/userGroups";
+import { UserGroups } from "@/screens/user-groups";
 
 export default function UserGroupsPage() {
   return <UserGroups />;
