+++ src/entities/permission/api/hooks/index.ts
@@ -2,3 +2,4 @@ export * from "./use-permission-search-v1";
 export * from "./use-fetch-by-id";
 export * from "./use-update";
 export * from "./use-create";
+export * from "./use-create-permission-grant-user-group";
