+++ src/entities/permission/api/hooks/use-create-permission-grant-user-group.ts
@@ -2,14 +2,16 @@ import { useCallback, useContext } from "react";
 
 import { ApiContext } from "@/shared/api";
 
-import { CreatePermissionGrantUserGroupRequestBody } from "../types/permission-grant";
+import { CreatePermissionGrantUserGroupRequestBody } from "../types";
 
 export const useCreatePermissionGrantUserGroup = () => {
   const api = useContext(ApiContext);
 
   const createPermissionGrantUserGroup = useCallback(
     async ({ body }: { body: CreatePermissionGrantUserGroupRequestBody }) => {
-      const { error } = await api.userGroup.createPermissionGrant({ body });
+      const { error } = await api.permission.createPermissionGrantUserGroup({
+        body,
+      });
 
       if (error) {
         throw error;
