+++ src/entities/permission/libs/constants.ts
@@ -1,4 +1,5 @@
 import { z } from "zod";
+
 import { FIRST_ID_EXTRACTOR } from "@/shared/libs";
 
 export const PERMISSION_SCHEMA = z.object({
@@ -10,4 +11,19 @@ export const PERMISSION_SCHEMA = z.object({
   key: z.string().min(1, "Key can not be empty"),
   name: z.string().min(1, "Name can not be empty"),
   description: z.string().optional(),
+});
+
+export const PERMISSION_GRANT_USER_GROUP_SCHEMA = z.object({
+  permissionSchemaId: z
+    .string()
+    .uuid("Permission schema ID must be a valid UUID")
+    .or(FIRST_ID_EXTRACTOR),
+  permissionId: z
+    .string()
+    .uuid("Permission ID must be a valid UUID")
+    .or(FIRST_ID_EXTRACTOR),
+  userGroupId: z
+    .string()
+    .uuid("User group ID must be a valid UUID")
+    .or(FIRST_ID_EXTRACTOR),
 });
