+++ src/entities/user-group/api/hooks/index.ts
@@ -1,3 +1,2 @@
 export * from "./use-permission-grant-user-group-search-v1";
 export * from "./use-search-v1";
-export * from "./use-create-permission-grant-user-group";
