+++ src/entities/user-group/components/resource-link/index.ts
