+++ src/entities/user-group/libs/hooks/index.ts
@@ -1,2 +1,2 @@
-export * from "./use-filters";
 export * from "./use-select-adapter";
+export * from "./use-filters";
