+++ src/entities/user-group/libs/hooks/use-select-adapter.tsx
