+++ src/entities/user-group/libs/index.ts
@@ -1,4 +1,3 @@
 export * from "./constants";
 export * from "./helpers";
 export * from "./hooks";
-export * from "./schema";
