--- src/entities/userGroup/libs/hooks/index.ts
@@ -1,2 +1,0 @@
-export * from "./useSelectAdapter";
-export * from "./useFilters";
