+++ src/screens/user-groups/index.ts
@@ -1,0 +1,1 @@
+export * from "./user-groups";
