--- src/screens/userGroups/index.ts
@@ -1,1 +1,0 @@
-export * from "./userGroups";
