+++ src/widgets/form-fields/featurer/hooks/useSelectAdapter.ts
@@ -1,3 +1,5 @@
+import { useMemo } from "react";
+
 import { useDatalistSelectAdapter } from "@/entities/datalist";
 import { FeaturerParamType } from "@/entities/featurer";
 import { useLinkSelectAdapter } from "@/entities/link";
@@ -12,9 +14,8 @@ import { useTwinClassSelectAdapter } from "@/entities/twin-class";
 import { useTwinClassFieldSelectAdapter } from "@/entities/twin-class-field";
 import { useTwinStatusSelectAdapter } from "@/entities/twin-status";
 import { useTwinFlowSchemaSelectAdapter } from "@/entities/twinFlowSchema";
-import { useUserGroupSelectAdapter } from "@/entities/userGroup";
-import { createFixedSelectAdapter, SelectAdapter } from "@/shared/libs";
-import { useMemo } from "react";
+import { useUserGroupSelectAdapter } from "@/entities/user-group";
+import { SelectAdapter, createFixedSelectAdapter } from "@/shared/libs";
 
 type FeaturerParamTypesSelectAdapter =
   | ReturnType<
